+++ src/App.js
@@ -2,7 +2,6 @@ import "./App.css";
 import HeaderTop from "./feature/HeaderTop/HeaderTop";
 import Nav from "./feature/Nav/Nav";
 import Highlight from "./feature/Highlight/Highlight";
-import News from "./feature/RecentNews/News";
 import Container from "./feature/Container/Container";
 
 
@@ -12,7 +11,6 @@ function App() {
             <HeaderTop />
             <Nav />
             <Highlight />
-            <News/>
             <Container />
         </div>
     );
