+++ src/feature/ArticleList/ArticleList.jsx
@@ -1,37 +1,35 @@
 import "./ArticleList.css";
+import React, { Component } from 'react'
 import ArticleItem from "../ArticleItem/ArticleItem.jsx";
 
-var articles = [
-    {
-        image: "https://static.zerochan.net/Genshin.Impact.full.3420945.jpg",
-        title: "11 Ngọc Châu đăng quang Hoa hậu Hoàn vũ Việt Nam 2022",
-        description: "11 Ukraine thừa nhận tư cách thành viên NATO không khả thi trong ngắn hạn, nhưng muốn liên minh thừa nhận Kiev là 'nền tảng' của an ninh châu Âu.",
-    },
-    {
-        image: "https://static.zerochan.net/Genshin.Impact.full.3420945.jpg",
-        title: "12 Ngọc Châu đăng quang Hoa hậu Hoàn vũ Việt Nam 2022",
-        description: "12 Ukraine thừa nhận tư cách thành viên NATO không khả thi trong ngắn hạn, nhưng muốn liên minh thừa nhận Kiev là 'nền tảng' của an ninh châu Âu.",
-    },
-    {
-        image: "https://static.zerochan.net/Genshin.Impact.full.3420945.jpg",
-        title: "13 Ngọc Châu đăng quang Hoa hậu Hoàn vũ Việt Nam 2022",
-        description: "13 Ukraine thừa nhận tư cách thành viên NATO không khả thi trong ngắn hạn, nhưng muốn liên minh thừa nhận Kiev là 'nền tảng' của an ninh châu Âu.",
-    },
-    {
-        image: "https://static.zerochan.net/Genshin.Impact.full.3420945.jpg",
-        title: "14 Ngọc Châu đăng quang Hoa hậu Hoàn vũ Việt Nam 2022",
-        description: "14 Ukraine thừa nhận tư cách thành viên NATO không khả thi trong ngắn hạn, nhưng muốn liên minh thừa nhận Kiev là 'nền tảng' của an ninh châu Âu.",
-    },
-];
+export default class ArticleList extends Component {
+    constructor() {
+        super();
+        this.state = {
+          items: []
+        };
+      }
 
-function ArticleList() {
-    return (
-        <div className="ArticleList">
-            {articles.map((article) => {
-                return <ArticleItem image={article.image} title={article.title} description={article.description} />;
-            })}
-        </div>
-    );
+    async componentDidMount() {
+        try {
+          const response = await fetch(
+            "https://api.rss2json.com/v1/api.json?rss_url=https%3A%2F%2Fvnexpress.net%2Frss%2Ftin-moi-nhat.rss"
+          );
+          const json = await response.json();
+          this.setState({ items: json.items });
+        } catch (error) {
+          console.log(error);
+        }
+    }
+
+      render() {
+        console.log(this.state);
+        return (
+          <div className="ArticleList">
+            {this.state.items.map(item => (
+                <ArticleItem key={item.guid} image={item.thumbnail} title={item.title} description={item.description.split("</a>")[1]} />
+            ))}
+          </div>
+        );
+      }
 }
-
-export default ArticleList;
